MA-22412-6 Add height alignment for YV12 format when calculate plane address offset
Below CTS case will fail because the YV12 format buffer plane address offset is
not match with CTS case expected one. Add height alignment when calculate the
size of each plane for some formats(DRM_FORMAT_YVU420 is the drm_format of YV12 format).

Failed CTS case:
MctsMediaDecoderTestCases android.media.decoder.cts.DecodeAccuracyTest#testGLViewLargerHeightDecodeAccuracy[4(c2.android.vp9.decoder_vp9_426x182)]

Reviewed-on: http://androidsource.nxp.com/project/26276

+++ display/gralloc/driver_utils.cpp
@@ -108,7 +108,7 @@ const format_info_t formats[] = {
         },
         {
                 .id = static_cast<int32_t>(PixelFormat::YV12),
-                .fourcc = static_cast<int32_t>(DRM_FORMAT_YVU420),
+                .fourcc = DRM_FORMAT_YVU420,
                 .modifier = DRM_FORMAT_MOD_LINEAR,
                 .is_rgb = false,
                 .is_yuv = true,
